views: Extract config_error view function.
